refactor: Create subfolder structure in components folder
