+++ web/src/components/sidebars/AnalyticsPanel.vue
