+++ web/src/components/sidebars/ChartsPanel.vue
