+++ web/src/components/sidebars/FloatingPanel.vue
