+++ web/src/components/sidebars/LayerStyle.vue
