+++ web/src/components/sidebars/LayersPanel.vue
