+++ web/src/components/sidebars/NodeAnimation.vue
